This programme will output the length of the longest continuous river segment that is not occupied by any animals.
Housing is a serious problem in Hong Kong. Someone wants to build a house along a river. Unfortunately, some parts of the river are occupied by animals and cannot be used. This programme will output the length of the longest continuous river segment that is not occupied by any animals.